ML-224 SVD decomposition does not produce original matrix
https://track.hpccsystems.com/i#browse/ML-224
https://track.hpccsystems.com/i#browse/ML-226

- Remove Lanczos optimization for calculation of eigen vectors and values

There's a number of optimizations that can be made in the ecl-ml matrix
libary whose implementation is based on sprase matrix representation.
For example, in the 'strict' mode of operation, matrix dimenssion should always
be exlicitely set, so that matrix dimensionality requirements can be enforced.
Once this is done, it should be possible to perform matrix thining and
value rounding to prevent unnecessarily carrying very small real numbers
through iterations, but without possibility of losing matrix
dimensionality during that process. This in turn will prevent error
propagation, and it will make it possible to re-introduce other matrix
operations optimizations, such as using Lanczos algorithm in calculation
of eigen vectors.



Signed-off-by: Edin Muharemagic <edin.muharemagic@lexisnexis.com>